Use the correct base image for builds
All our builds are basing on the openshift/origin-base and it turns
out that each release has a base image that should be used for builds
for that release - e.g. for 4.3 it is ocp/4.3:base

Just like ci-operator initiated builds, set this up by using the
'from' config for the docker build strategy. See:

https://docs.openshift.com/container-platform/4.3/builds/build-strategies.html#builds-strategy-docker-from-image_build-strategies

The equivalent ci-operator config is:

  base_images:
    base:
      name: "4.3"
      namespace: ocp
      tag: base

You can tell this is working because the log says:

  Pulling image docker-registry.default.svc:5000/ocp/4.3@sha256:8f4728ddd27d43b8afc0b46f804db14dc2df87c9ddcc95b85a62e64fb03c4d2f

rather than:

  Pulling image openshift/origin-base